WebSocket: Header JWT authorization
JWT added to the header of CONNECT stomp message
so we can retrieve it at ChannelInterceptor and set User
of this message if JWT is valid. To do this we use the same
mechanism, which was already implemented for REST
(refactored to the new AuthUserService in this patch also).
So it works the same way, as for REST athorization.

Later, in message hadling method of ChatController, we can
get authenticated user from message headers.

As far as I understand for now, this is the most OK way to
do this. See the official Spring documentation
https://github.com/spring-projects/spring-framework/blob/main/src/docs/asciidoc/web/websocket.adoc#token-authentication