MessageEdit: Backend and websocket
Here we added saving of edited message
to our database. We also added new
column to ChatMessage table to save
previous message text (before editing).
This allows us to differentiate between
pristine and edited message on frontend
and, in future, to show edit diff to curious
user. For now we think that it is enough to
store only one message snapshot (previous).
We also added websocket GLOBAL_EVENT
to change loaded message in other users
chats.