ReplyToMessage: Scroll to previous message
Here we added ability for user to go to replied
message. To do this we load all messages after
requested message date (including it) and then
scroll chat so that message is in view.
We highlight that message and also show "Back"
button, allowing user to return to previous
message.
That should work in case of reply chain too,
but testing is needed.